src/requirements.txt
Django==1.10.3
django-filter==0.15.3
djangorestframework==3.5.3
gunicorn==19.6.0
Markdown==2.6.7
mysql==0.0.1
MySQL-python==1.2.5
PyYAML==3.12
